aopcs_lambda/target/aws/requirements.txt
aws-cdk-lib==2.194.0
constructs>=10.4.2,<11.0.0
